[02旅游景点智能推荐平台] feat：添加景观组件并增强景观细节功能
- 引入ScenicCardGrid、ScenicDetailGallery、ScenicDetailHeader、ScenicListBanner、ScenicListCard、ScenicListFilterBar、ScenicSidePanel和ScenicSimilarSection组件，以提升UI。

- 在AdminScenicSpotController中实现了管理员详情端点，用于获取景观详情。

- 将ScenicUpdateDTO中的ratingScore类型从BigDecimal更改为Double。

- 通过添加adminDetail方法增强ScenicSpotService，以实现详细景观检索。

- 为ScenicDetailVO添加tagIds字段，以实现更好的标签管理。

- 为ScenicFilterOptionsVO更新了描述性注释，以增强清晰度。

+++ projects/02travel_recommend_system/travel-web/src/main/java/com/travel/advisor/dto/scenic/ScenicUpdateDTO.java
@@ -5,6 +5,9 @@ import lombok.Data;
 import java.math.BigDecimal;
 import java.util.List;
 
+/**
+ * 景点更新 DTO
+ */
 @Data
 public class ScenicUpdateDTO {
 
@@ -34,7 +37,7 @@ public class ScenicUpdateDTO {
 
     private String category;
 
-    private BigDecimal ratingScore;
+    private Double ratingScore;
 
     private String bestSeason;
